+++ package.json
@@ -14,6 +14,7 @@
     "graphql": "^15.5.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-router-dom": "5.3.0",
     "react-scripts": "4.0.3",
     "web-vitals": "1.1.2"
   },
