+++ src/Index.tsx
@@ -6,6 +6,7 @@ import reportWebVitals from './ReportWebVitals';
 import { ApolloProvider,
   ApolloClient,
   InMemoryCache } from "@apollo/client";
+import { BrowserRouter } from 'react-router-dom';
 
 const client = new ApolloClient({
   uri: 'https://tq-template-server-sample.herokuapp.com/graphql',
@@ -15,7 +16,7 @@ const client = new ApolloClient({
 ReactDOM.render(
   <React.StrictMode>
     <ApolloProvider client={client}>
-      <App />
+        <App />
     </ApolloProvider>
   </React.StrictMode>,
   document.getElementById('root')
