+++ src/NewPage.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export default function NewPage(){
+    return(
+    <div>
+        <h1>Nova Página</h1>     
+    </div>
+    )
+  }
+
